# flate: Improve huffman generation speed ~5% faster